An experiment to move qiskit.extensions to a different directory
