+++ qiskit/__init__.py
@@ -32,7 +32,6 @@ from qiskit import user_config as _user_config
 
 # The qiskit.extensions.x imports needs to be placed here due to the
 # mechanism for adding gates dynamically.
-import qiskit.extensions
 import qiskit.circuit.measure
 import qiskit.circuit.reset
