+++ qiskit/circuit/library/generalized_gates/hamiltonian_gate.py
