+++ qiskit/circuit/library/quantum_initializer/diagonal.py
