+++ qiskit/circuit/library/quantum_initializer/isometry.py
@@ -29,8 +29,8 @@ from qiskit.circuit.quantumcircuit import QuantumCircuit
 from qiskit.circuit.quantumregister import QuantumRegister
 from qiskit.exceptions import QiskitError
 from qiskit.quantum_info.operators.predicates import is_isometry
-from qiskit.extensions.quantum_initializer.uc import UCGate
-from qiskit.extensions.quantum_initializer.mcg_up_to_diagonal import MCGupDiag
+from qiskit.circuit.library.quantum_initializer.uc import UCGate
+from qiskit.circuit.library.quantum_initializer.mcg_up_to_diagonal import MCGupDiag
 
 _EPS = 1e-10  # global variable used to chop very small numbers to zero
