+++ qiskit/circuit/library/quantum_initializer/squ.py
