+++ qiskit/circuit/library/quantum_initializer/ucrx.py
@@ -26,7 +26,7 @@ import math
 from qiskit.circuit.quantumcircuit import QuantumCircuit
 from qiskit.circuit.quantumregister import QuantumRegister
 from qiskit.exceptions import QiskitError
-from qiskit.extensions.quantum_initializer.uc_pauli_rot import UCPauliRotGate
+from qiskit.circuit.library.quantum_initializer.uc_pauli_rot import UCPauliRotGate
 
 
 class UCRXGate(UCPauliRotGate):
