+++ qiskit/circuit/library/simulator/snapshot.py
