+++ qiskit/transpiler/passes/optimization/hoare_opt.py
@@ -14,12 +14,11 @@
 from qiskit.transpiler.basepasses import TransformationPass
 from qiskit.circuit import QuantumRegister, ControlledGate, Gate
 from qiskit.dagcircuit import DAGCircuit
-from qiskit.extensions.unitary import UnitaryGate
+from qiskit.circuit.library.generalized_gates.unitary import UnitaryGate
 from qiskit.quantum_info.operators.predicates import matrix_equal
 from qiskit.transpiler.exceptions import TranspilerError
 from qiskit.circuit.exceptions import CircuitError
 from qiskit.circuit.library.standard_gates import CZGate, CU1Gate, MCU1Gate
-from . import _gate_extension  # pylint: disable=unused-import
 
 try:
     from z3 import And, Or, Not, Implies, Solver, Bool, unsat
