+++ test/ipynb/mpl/circuit/test_circuit_matplotlib_drawer.py
@@ -294,7 +294,6 @@ class TestMatplotlibDrawer(QiskitTestCase):
         circuit.h(q[1])
 
         # this import appears to be unused, but is actually needed to get snapshot instruction
-        import qiskit.extensions.simulator  # pylint: disable=unused-import
 
         circuit.snapshot("1")
