+++ test/python/circuit/test_uc.py
@@ -22,7 +22,7 @@ import itertools
 import numpy as np
 from scipy.linalg import block_diag
 
-from qiskit.extensions.quantum_initializer.uc import UCGate
+from qiskit.circuit.library.quantum_initializer.uc import UCGate
 
 from qiskit import QuantumCircuit, QuantumRegister, BasicAer, execute
 from qiskit.test import QiskitTestCase
