+++ test/python/visualization/test_circuit_text_drawer.py
@@ -28,7 +28,7 @@ from qiskit.transpiler import Layout
 from qiskit.visualization import text as elements
 from qiskit.visualization.circuit_visualization import _text_circuit_drawer
 from qiskit.extensions import UnitaryGate, HamiltonianGate
-from qiskit.extensions.quantum_initializer import UCGate
+from qiskit.circuit.library.quantum_initializer import UCGate
 from qiskit.circuit.library import (
     HGate,
     U2Gate,
